Fixed signed/unsigned comparison.

+++ src/SampleRegulator.cpp
@@ -192,7 +192,7 @@ namespace geopm
             }
         }
 
-        for (int i = 0; i < num_domain; ++i) {
+        for (size_t i = 0; i < num_domain; ++i) {
             telemetry[i].region_id = m_region_id_prev;
             telemetry[i].timestamp = m_aligned_time;
             memcpy(telemetry[i].signal,
